DataScience/Jupyter Notebook/FireDetect/3DFire/fire-loc3d-coplt/requirements.txt
numpy>=1.18.0
opencv-python>=4.5.0
torch>=1.8.0
torchvision>=0.9.0
PyYAML>=5.4.0
matplotlib>=3.3.0
scipy>=1.6.0
pillow>=8.0.0
tqdm>=4.56.0
open3d>=0.13.0
pandas>=1.2.0
